Extracted Ghost concerns into a Ghost Manager
